Migrated ApiPage to typescript
