--- src/views/api/shared/ApiPage/index.js
@@ -1,1 +1,0 @@
-export { default } from './ApiPage';
