+++ src/views/api/shared/ApiPage/index.ts
@@ -1,0 +1,1 @@
+export * from './ApiPage';
